+++ cobib/commands/delete.py
@@ -1,12 +1,15 @@
 """CoBib delete command."""
 
 import argparse
+import logging
 import os
 import sys
 
 from cobib.config import CONFIG
 from cobib.database import read_database
 from .base_command import ArgumentParser, Command
+
+LOGGER = logging.getLogger(__name__)
 
 
 class DeleteCommand(Command):
@@ -21,6 +24,7 @@ class DeleteCommand(Command):
 
         Args: See base class.
         """
+        LOGGER.debug('Starting Delete command.')
         parser = ArgumentParser(prog="delete", description="Delete subcommand parser.")
         parser.add_argument("label", type=str, help="label of the entry")
 
@@ -42,10 +46,12 @@ class DeleteCommand(Command):
         buffer = []
         for line in lines:
             if line.startswith(largs.label):
+                LOGGER.debug('Entry "%s" found. Starting to remove lines.', largs.label)
                 entry_to_be_deleted = True
                 buffer.pop()
                 continue
             if entry_to_be_deleted and line.startswith('...'):
+                LOGGER.debug('Reached end of entry "%s".', largs.label)
                 entry_to_be_deleted = False
                 continue
             if not entry_to_be_deleted:
@@ -57,13 +63,14 @@ class DeleteCommand(Command):
     @staticmethod
     def tui(tui):
         """See base class."""
+        LOGGER.debug('Delete command triggered from TUI.')
         # get current label
         label, _ = tui.get_current_label()
         # delete selected entry
         DeleteCommand().execute([label])
-        # update bibliography data
-        read_database(fresh=True)
         # update database list
+        LOGGER.debug('Updating list after Delete command.')
+        read_database(fresh=True)
         tui.update_list()
         # if cursor line is below buffer height, move it one line back up
         if tui.current_line >= tui.buffer.height: