+++ cobib/commands/open.py
@@ -1,11 +1,14 @@
 """CoBib open command."""
 
 import argparse
+import logging
 import sys
 from subprocess import Popen
 
 from cobib.config import CONFIG
 from .base_command import ArgumentParser, Command
+
+LOGGER = logging.getLogger(__name__)
 
 
 class OpenCommand(Command):
@@ -13,13 +16,14 @@ class OpenCommand(Command):
 
     name = 'open'
 
-    def execute(self, args, out=sys.stdout):
+    def execute(self, args, out=sys.stderr):
         """Open file from entry.
 
         Opens the associated file of an entry with xdg-open.
 
         Args: See base class.
         """
+        LOGGER.debug('Starting Open command.')
         parser = ArgumentParser(prog="open", description="Open subcommand parser.")
         parser.add_argument("label", type=str, help="label of the entry")
 
@@ -36,26 +40,31 @@ class OpenCommand(Command):
         try:
             entry = CONFIG.config['BIB_DATA'][largs.label]
             if 'file' not in entry.data.keys() or entry.data['file'] is None:
-                error = "Error: There is no file associated with this entry."
+                msg = "Error: There is no file associated with this entry."
+                LOGGER.error(msg)
                 if out is None:
                     # called from TUI
-                    return error
-                print(error, file=out)
+                    return msg
+                print('Error: ' + msg, file=out)
                 sys.exit(1)
             opener = None
             opener = CONFIG.config['DATABASE'].get('open')
             try:
+                LOGGER.error('Opening "%s" with %s.', entry.data['file'], opener)
                 Popen([opener, entry.data['file']])
             except FileNotFoundError:
                 pass
         except KeyError:
             print("Error: No entry with the label '{}' could be found.".format(largs.label))
+            LOGGER.error(msg)
+            print(msg, file=out)
 
         return None
 
     @staticmethod
     def tui(tui):
         """See base class."""
+        LOGGER.debug('Open command triggered from TUI.')
         # get current label
         label, _ = tui.get_current_label()
         # populate buffer with entry data