+++ test/__init__.py
@@ -1,1 +1,4 @@
 """CoBib test module."""
+import logging
+
+logging.getLogger().setLevel(logging.DEBUG)